requirements.txt
streamlit
pandas
ultralytics
opencv-python
pillow
numpy
torch
torchvision
firebase-admin
requests
flask
flask-cors
